Rename redis -> redis_storage
